% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pakiet.R
\name{plot.przyrosty}
\alias{plot.przyrosty}
\title{Przyrosty Plot}
\usage{
\method{plot}{przyrosty}(x, ...)
}
\arguments{
\item{x}{wynik roznicowania}

\item{...}{inne argumenty}
}
\description{
Funkcja zwraca wykres dla obliczonych wczesniej przyrostow dla danych.
}
\author{
Karolina Dudzinska, Bartosz Nguyen Van
}
