% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pakiet.R
\name{print.przyrosty}
\alias{print.przyrosty}
\title{Przyrosty Print}
\usage{
\method{print}{przyrosty}(x, ...)
}
\arguments{
\item{x}{wynik roznicowania}

\item{...}{inne argumenty}
}
\description{
Funkcja wypisuje obliczone wczesniej przyrosty dla danych: standardowe lub logarytmiczne.
}
\author{
Karolina Dudzinska, Bartosz Nguyen Van
}
