% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pakiet.R
\name{przyrosty.default}
\alias{przyrosty.default}
\title{Przyrosty}
\usage{
\method{przyrosty}{default}(x, lag = 1L, roznice = 1L, logarytm = FALSE, ...)
}
\arguments{
\item{x}{macierz lub numeryczny wektor danych zwierajacy wartosci do zroznicowania}

\item{lag}{wartosc numeryczna typu integer okreslajaca liczne opoznien}

\item{roznice}{wartosc numeryczna typu integer okreslajaca kolejnosc roznicowania, ile razy roznicowanie ma byc przeprowadzone na danych}

\item{logarytm}{wartosc typu bool (TRUE lub FALSE), okreslajaca czy dane maja zostac zlogarytmizowane, defaultowo dane nie sa przeksztalcane na logarytm}

\item{...}{inne argumenty}
}
\value{
Jesli x jest wektorem o dlugosci n i roznica = 1, to obliczony wynik jest rowny kolejnym roznicom x[(1+lag):n] - x[1:(n-lag)].
Jezeli roznica jest wieksza od 1, to algorytm ten jest stosowany rekurencyjnie do x.
Zwracana wartoscia jest wektor krotszy od x.
Jezeli x jest macierza, to operacje roznicy sa wykonywane na kazdej kolumnie osobno.
}
\description{
Funkcja zwraca przyrosty dla danych: standardowe lub logarytmiczne.
}
\author{
Karolina Dudzinska, Bartosz Nguyen Van
}
