% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pakiet.R
\name{summary.przyrosty}
\alias{summary.przyrosty}
\title{Przyrosty Summary}
\usage{
\method{summary}{przyrosty}(object, ...)
}
\arguments{
\item{object}{wynik roznicowania}

\item{...}{inne argumenty}
}
\description{
Funkcja zwraca podstawowe statystyki opisowe dla obliczonych wczesniej przyrostow.
}
\author{
Karolina Dudzinska, Bartosz Nguyen Van
}
